some data types corrected
size_t is the unsigned variant of int and in case of Arduino of short, too.

+++ cnc_ctrl_v1/GCode.cpp
@@ -461,7 +461,7 @@ void  sanitizeCommandString(String& cmdString){
     */
     
     byte line_flags = 0;
-    short pos = 0;
+    size_t pos = 0;
     
     while (cmdString.length() > pos){
         if (line_flags) {
@@ -519,8 +519,8 @@ byte  interpretCommandString(String& cmdString){
     
     */
     
-    int firstG;  
-    int secondG;
+    size_t firstG;  
+    size_t secondG;
 
     if (cmdString.length() > 0) {
         if (cmdString[0] == '$') {
